# auth-service/src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ResponseData": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"statusCode",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"isVerified": {
						"type": "boolean"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"resetPasswordTokens": {
						"items": {
							"$ref": "#/components/schemas/ResetPasswordToken"
						},
						"type": "array"
					},
					"refreshTokens": {
						"items": {
							"$ref": "#/components/schemas/RefreshToken"
						},
						"type": "array"
					},
					"emailVerificationTokens": {
						"items": {
							"$ref": "#/components/schemas/EmailVerificationToken"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"email",
					"password",
					"isActive",
					"isVerified",
					"role",
					"resetPasswordTokens",
					"refreshTokens",
					"emailVerificationTokens"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Role": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"users": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"users"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResetPasswordToken": {
				"properties": {
					"id": {
						"type": "string"
					},
					"token_hash": {
						"type": "string"
					},
					"isUsed": {
						"type": "boolean"
					},
					"expiresAt": {
						"type": "string",
						"format": "date-time"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					}
				},
				"required": [
					"id",
					"token_hash",
					"isUsed",
					"expiresAt",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RefreshToken": {
				"properties": {
					"id": {
						"type": "string"
					},
					"token_hash": {
						"type": "string"
					},
					"isUsed": {
						"type": "boolean"
					},
					"expiresAt": {
						"type": "string",
						"format": "date-time"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					}
				},
				"required": [
					"id",
					"token_hash",
					"isUsed",
					"expiresAt",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EmailVerificationToken": {
				"properties": {
					"id": {
						"type": "string"
					},
					"token_hash": {
						"type": "string"
					},
					"expiresAt": {
						"type": "string",
						"format": "date-time"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					}
				},
				"required": [
					"id",
					"token_hash",
					"expiresAt",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "auth-service",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users/register": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"accessToken": {
											"type": "string"
										},
										"role": {
											"type": "string"
										},
										"email": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"accessToken",
										"role",
										"email",
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"role": {
										"type": "string"
									},
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"role",
									"password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/users/login": {
			"post": {
				"operationId": "LoginUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"accessToken": {
											"type": "string"
										},
										"user": {
											"properties": {
												"role": {
													"type": "string"
												},
												"isVerified": {
													"type": "boolean"
												},
												"isActive": {
													"type": "boolean"
												},
												"email": {
													"type": "string"
												},
												"id": {
													"type": "string"
												}
											},
											"required": [
												"role",
												"isVerified",
												"isActive",
												"email",
												"id"
											],
											"type": "object"
										}
									},
									"required": [
										"accessToken",
										"user"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/users/verify-email": {
			"post": {
				"operationId": "VerifyEmail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"token": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"token",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/delete/{id}": {
			"get": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/reset-password": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"token": {
										"type": "string"
									},
									"newPassword": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"token",
									"newPassword",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/users/forgot-password": {
			"post": {
				"operationId": "ForgotPassword",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"email": {
										"type": "string"
									}
								},
								"required": [
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/users/initiate-email-verification": {
			"post": {
				"operationId": "InitiateEmailVerification",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"email": {
										"type": "string"
									}
								},
								"required": [
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}